Enhanced Ankimon (solved bugs + new features) (#311)
* New pokedex.html

Enhanced pokedex view

* Enhanced pokedex_obj.py

To manage the changes on `pokedex_obj.py`

* Created pokemon_names.json

`pokemon_names.json` is made to manage the changes on `pokedex.html`

* Update pokedex.html

* New pokedex.html

* Update pokedex_obj.py

* Updated reviewer_obj.py

Added a function to change colors while on light and dark mode.

* simple typo targeting main_pokemon.x instead of main_pokemon.x*p*

* New pokedex.html

Added a part so that every text (except buttons) is white while in dark mode.

* pokedex_obj.py

Updated pokedex python code, so far free of bugs, made a one line adjustment.

* pokedex.html

This latest version has a radiant glow behind each Pokémon when hovering, and the names now display in white while in dark mode. I believe this is final, being a stable and usable version.

* pokemon_details.py

* ankimon_tracker.py

Solved an issue that the tracker didn't adapt to light/dark mode

---------

Co-authored-by: h0tp-ftw <noddy.manan@gmail.com>